IVC: extracting PlonkishWitness to reuse for prover
Failing attempt as it requires to index by the Column. which is not in the same
crate.
