+++ ivc/src/ivc/mod.rs
@@ -1,5 +1,6 @@
 pub mod columns;
 pub mod constraints;
+pub mod folding;
 pub mod helpers;
 pub mod interpreter;
 pub mod lookups;
